w3c-reports/12092014/gtk/runned-tests/XMLHttpRequest/send-accept-language-expected.txt


FAIL Send "sensible" default value, whatever that means assert_regexp_match: expected object "/accept-language:\s.+/" but got ""(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_regexp_match@http://localhost:8800/resources/testharness.js:504:15
http://localhost:8800/XMLHttpRequest/send-accept-language.htm:16:28
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/XMLHttpRequest/send-accept-language.htm:12:11)
PASS XMLHttpRequest: send() - Accept-Language 
